pyproject.toml
[tool.poetry]
name = "coursework-7"
version = "0.1.0"
description = ""
authors = ["Futeki <Povizun@bk.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.6"
psycopg2-binary = "^2.9.9"
ipython = "^8.25.0"
pytils = "^0.4.1"
python-dotenv = "^1.0.1"
djangorestframework = "^3.15.2"
black = "^24.4.2"
isort = "^5.13.2"
django-filter = "^24.2"
djangorestframework-simplejwt = "^5.3.1"
coverage = "^7.5.4"
drf-yasg = "^1.21.7"
celery = "^5.4.0"
eventlet = "^0.36.1"
redis = "^5.0.7"
django-celery-beat = "^2.6.0"
pillow = "^10.3.0"
django-cors-headers = "^4.4.0"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
